Update Readme and refactor utilty methods into utlity file
