+++ README.md
@@ -96,3 +96,128 @@ In case you are missing an important tutorial, feel free to request it via a Git
 ### API Reference
 
 Public classes / methods / properties are annotated with Docstrings. The compiled API Reference can be found under [Documentation -> API Reference](https://parallel-domain.github.io/pd-sdk/)
+
+## Compare DGP testing tool
+
+### Overview
+
+A new testing tool was added under `compare_dgp_scene.py`. The goal of this tool is to take two DGP scenes that should 
+be identical and perform a frame and sensor level comparison for the annotation streams that are currently supported. 
+This tool was originally created to compare datasets produced in the current pipeline and those generated by the LEC 
+rendered pipeline.
+
+The tool is still currently a work in progress and only supports HTML and text based error reporting. Development of 
+this tool will continue alongside the LEC as more annotation streams are supported
+
+### Supported annotation streams
+
+* Ontology
+* Camera Model
+* RGB
+* Instance Mask
+* Semantic Mask
+* Depth
+* Bounding Box 2D
+* Bounding Box 3D
+
+### Installation
+
+Install dev dependencies.
+
+```bash
+$ pip install -e .[dev]
+```
+
+### Running
+
+Help docs can be generated by running the following:
+
+```bash
+$ python3 compare_dgp_scene.py --help
+```
+
+Running the tool on a test dataset
+
+```bash
+$ python3 compare_dgp_scene.py -k "[f0-" --test-dataset /pipelineData/goku_dgp_dataset/dataset --target-dataset /pipelineData/goku_test/keystone/dataset --test-scene scene_000000 --output-dir "/pd-sdk/images" --html=test-results.html --capture=tee-sys
+```
+
+### Configuration parameters and test details
+
+This section gives a brief outline of each test and it's associated threshold parameters.
+
+#### Ontology
+
+Do a 1-1 comparison of all the ontologies listed in each ontology files.
+Has no threshold parameters and expects files to be exactly the same.
+
+#### Camera Model
+
+Do a 1-1 comparison of all the ontologies listed in each ontology files.
+Has no threshold parameters and expects files to be exactly the same.
+
+#### RGB
+
+Compare RGB images at a pixel level to see if they are identical.
+
+* RGB_PIXEL_DIFF_THRESHOLD = 5 
+  * Used to set the maximum difference between images
+  * Currently quite high due to render differences for tress and traffic lights
+
+#### Instance Mask
+
+Compare Instance images at a pixel level to see if they are identical.
+
+* INST_SEG_PIXEL_DIFF_THRESHOLD = 2
+  * Used to set the maximum difference between images
+* MIN_INSTANCED_OBJECT_PERCENTAGE_OVERLAP = 99.5
+  * When comparing two instances the minium number of pixels that have a 1-1 match
+  * Set to be low to catch as many errors as possible as noise here is low
+
+#### Semantic Mask
+
+Compare Semantic images at a pixel level to see if they are identical.
+
+* SEM_SEG_PIXEL_DIFF_THRESHOLD = 2
+  * Used to set the maximum difference between images
+
+#### Depth
+
+Compare Depth images at a pixel level to see if they are identical.
+
+* DEPTH_PIXEL_DIFF_THRESHOLD = 2
+  * Used to set the maximum difference between depth values 
+
+#### Bounding Box 2D
+
+Compare 2d bounding boxes. Attempted to find a 1-1 matched for each bounding boxes
+For matched bounding boxes check the following
+* Have matching attributes
+* Are within x size range
+
+Current matching algorithm is intersection over union
+
+* MIN_2D_BBOX_BOX_SIZE = 10
+ * Size cut off for bounding boxes
+* MIN_2D_BBOX_IOU = 0.75
+ * Minimum IOU value
+* MIN_PIXEL_SIZE_FOR_STRICTNESS = 150
+ * Allow more variability for small boxes to improve matching
+* MAX_2D_BBOX_PERCENT_SIZE_DIFF = 5
+ * Max size difference
+
+#### Bounding Box 3D
+
+Compare 3d bounding boxes. Attempted to find a 1-1 matched for each bounding boxes
+For matched bounding boxes check the following
+* Have matching attributes
+* Are within x size range
+
+Current matching algorithm sums the pythagorean theorem distance between each of the 3d boxes vertices.
+
+* MIN_3D_BBOX_VOLUME = 0.1
+ * Lower bound for 3d box volume
+* MIN_3D_BBOX_BETWEEN_VERTS_DISTANCE = 200
+ * Min distance between vertices 
+* MAX_3D_BBOX_PERCENT_VOLUME_DIFF = 5
+  * Max size difference
