+++ compare_dgp_scene_utils.py
